Update 2016-05-18_Setting_Up_LAMP.md
I added in more detail about why I chose to use Rufus as opposed to win32diskimager and fixed the bit about downloading a headless image. However, I wrote this guide with the intention of it being a precursor to installing an ownCloud server, as well. This requires a full installation of php5 as it makes use of some extensions to make it run faster on a Pi. That is why I suggested the full php5 package as opposed to just the Apache module.