Quicklinks1 (#15)
* Create 2016-05-18_Setting_Up_LAMP.md (#12)

* Create 2016-05-18_Setting_Up_LAMP.md

* Update 2016-05-18_Setting_Up_LAMP.md

I added in more detail about why I chose to use Rufus as opposed to win32diskimager and fixed the bit about downloading a headless image. However, I wrote this guide with the intention of it being a precursor to installing an ownCloud server, as well. This requires a full installation of php5 as it makes use of some extensions to make it run faster on a Pi. That is why I suggested the full php5 package as opposed to just the Apache module.

* Update 2016-05-18_Setting_Up_LAMP.md

* Update and rename 2016-05-18_Setting_Up_LAMP.md to 2016-05-18_Setting_Up_LAMP_Part_1.md

Changed the post to reflect it being part of a two-part series. Removed SSL information, information about port forwarding and domain purchasing, and generally just keeping it off of the internet, period.

* Rename 2016-05-18_Setting_Up_LAMP_Part_1.md to 2016-05-25_Setting_Up_LAMP_Part_1.md

* Bash history initial commit

* minor style changes

* fix bullets

* initial draft

* initial draft